Responsiveness
CSS:
In order to make it responsive a fixed width is not possible. Using a
position absolute and left and right works.

JS:
!!Haven't compiled a minified version!!
For convenience "/src/" went into "filePath"
Setting addressBarWidth and addressTxtWidth was redundant as far as I
could tell - in this way, might needs more testing.
Set some ratios upon resize.

HTML:
A little demo page that I used to test it with.